refactor(registry): split into types and versions api
