+++ registry/version_api_test.go
